+++ src/main/java/net/buildtheearth/buildteamtools/modules/generator/components/rail/RailScripts.java
@@ -1,15 +1,14 @@
 package net.buildtheearth.buildteamtools.modules.generator.components.rail;
 
-import com.alpsbte.alpslib.utils.GeneratorUtils;
 import com.sk89q.worldedit.EditSession;
 import com.sk89q.worldedit.WorldEdit;
 import com.sk89q.worldedit.bukkit.BukkitAdapter;
-import com.sk89q.worldedit.math.BlockVector2;
 import com.sk89q.worldedit.math.BlockVector3;
-import com.sk89q.worldedit.regions.CuboidRegion;
-import com.sk89q.worldedit.regions.Polygonal2DRegion;
 import net.buildtheearth.buildteamtools.BuildTeamTools;
 import net.buildtheearth.buildteamtools.modules.generator.GeneratorModule;
+import net.buildtheearth.buildteamtools.modules.generator.components.rail.path.RailPath;
+import net.buildtheearth.buildteamtools.modules.generator.components.rail.path.RailPathBuilder;
+import net.buildtheearth.buildteamtools.modules.generator.components.rail.selection.RailSelectionPointReader;
 import net.buildtheearth.buildteamtools.modules.generator.model.GeneratorComponent;
 import net.buildtheearth.buildteamtools.modules.generator.model.GeneratorType;
 import net.buildtheearth.buildteamtools.modules.generator.model.History;
@@ -54,285 +53,28 @@ public class RailScripts extends Script {
                 return;
             }
 
-            List<Vector> centerPath = createCenterPath(controlPoints);
+            RailPath railPath = new RailPathBuilder().build(controlPoints);
 
-            if (centerPath.size() < 2) {
+            if (!railPath.isValid()) {
                 getPlayer().sendMessage("§cRail Generator could not derive a valid path from this selection.");
                 return;
             }
 
-            Bukkit.getScheduler().runTask(BuildTeamTools.getInstance(), () -> placeSampleTrack(centerPath));
+            Bukkit.getScheduler().runTask(BuildTeamTools.getInstance(), () -> placeSampleTrack(railPath));
         } catch (Exception exception) {
-            getPlayer().sendMessage("§cRail Generator failed while reading the WorldEdit selection.");
+            getPlayer().sendMessage("§cRail Generator failed while generating from the WorldEdit selection.");
             exception.printStackTrace();
         }
     }
 
     private List<Vector> getControlPoints() {
-        if (getRegion() instanceof CuboidRegion cuboidRegion)
-            return getCuboidControlPoints(cuboidRegion);
-
-        if (getRegion() instanceof Polygonal2DRegion polygonalRegion)
-            return getPolygonalControlPoints(polygonalRegion);
-
-        List<Vector> points = GeneratorUtils.getSelectionPointsFromRegion(getRegion());
-
-        if (points == null || points.size() < 2)
-            return new ArrayList<>();
-
-        List<Vector> blockPoints = new ArrayList<>();
-
-        for (Vector point : points)
-            blockPoints.add(toBlockVector(point));
-
-        return orderPointsAsPath(blockPoints);
+        RailSelectionPointReader reader = new RailSelectionPointReader(getPlayer(), getRegion());
+        return reader.readControlPoints();
     }
 
-    private List<Vector> getCuboidControlPoints(CuboidRegion cuboidRegion) {
-        List<Vector> points = new ArrayList<>();
+    private void placeSampleTrack(RailPath railPath) {
+        List<Vector> centerPath = railPath.getCenterPath();
 
-        BlockVector3 pos1 = cuboidRegion.getPos1();
-        BlockVector3 pos2 = cuboidRegion.getPos2();
-
-        Vector start = new Vector(pos1.x(), pos1.y(), pos1.z());
-        Vector end = new Vector(pos2.x(), pos2.y(), pos2.z());
-
-        if (sameBlock(start, end)) {
-            start = new Vector(
-                    cuboidRegion.getMinimumPoint().x(),
-                    cuboidRegion.getMinimumPoint().y(),
-                    cuboidRegion.getMinimumPoint().z()
-            );
-
-            end = new Vector(
-                    cuboidRegion.getMaximumPoint().x(),
-                    cuboidRegion.getMinimumPoint().y(),
-                    cuboidRegion.getMaximumPoint().z()
-            );
-        }
-
-        points.add(start);
-        points.add(end);
-
-        return points;
-    }
-
-    private List<Vector> getPolygonalControlPoints(Polygonal2DRegion polygonalRegion) {
-        List<Vector> points = new ArrayList<>();
-
-        int minY = polygonalRegion.getMinimumY();
-        int maxY = polygonalRegion.getMaximumY();
-
-        for (BlockVector2 point : polygonalRegion.getPoints()) {
-            int y = findBestYForPolygonPoint(point.x(), point.z(), minY, maxY);
-            points.add(new Vector(point.x(), y, point.z()));
-        }
-
-        if (points.size() < 2)
-            return points;
-
-        return removeOnlyConsecutiveDuplicates(points);
-    }
-
-    private int findBestYForPolygonPoint(int x, int z, int minY, int maxY) {
-        org.bukkit.World world = getPlayer().getWorld();
-
-        int safeMinY = Math.max(world.getMinHeight(), Math.min(minY, maxY));
-        int safeMaxY = Math.min(world.getMaxHeight() - 1, Math.max(minY, maxY));
-
-        for (int y = safeMaxY; y >= safeMinY; y--) {
-            Material material = world.getBlockAt(x, y, z).getType();
-
-            if (material.isSolid())
-                return y;
-        }
-
-        return safeMaxY;
-    }
-
-    private List<Vector> orderPointsAsPath(List<Vector> points) {
-        List<Vector> remaining = new ArrayList<>();
-        List<Vector> ordered = new ArrayList<>();
-
-        for (Vector point : points) {
-            if (!containsBlock(remaining, point))
-                remaining.add(point);
-        }
-
-        if (remaining.isEmpty())
-            return ordered;
-
-        Vector current = findStartPoint(remaining);
-        ordered.add(current);
-        remaining.remove(current);
-
-        while (!remaining.isEmpty()) {
-            Vector next = findNearestPoint(current, remaining);
-            ordered.add(next);
-            remaining.remove(next);
-            current = next;
-        }
-
-        return ordered;
-    }
-
-    private Vector findStartPoint(List<Vector> points) {
-        Vector best = points.get(0);
-
-        for (Vector point : points) {
-            if (point.getBlockX() < best.getBlockX()) {
-                best = point;
-                continue;
-            }
-
-            if (point.getBlockX() == best.getBlockX() && point.getBlockZ() < best.getBlockZ())
-                best = point;
-        }
-
-        return best;
-    }
-
-    private Vector findNearestPoint(Vector from, List<Vector> points) {
-        Vector best = points.get(0);
-        double bestDistance = distanceSquared2D(from, best);
-
-        for (Vector point : points) {
-            double distance = distanceSquared2D(from, point);
-
-            if (distance < bestDistance) {
-                best = point;
-                bestDistance = distance;
-            }
-        }
-
-        return best;
-    }
-
-    private double distanceSquared2D(Vector a, Vector b) {
-        double dx = a.getBlockX() - b.getBlockX();
-        double dz = a.getBlockZ() - b.getBlockZ();
-
-        return dx * dx + dz * dz;
-    }
-
-    private boolean containsBlock(List<Vector> points, Vector target) {
-        for (Vector point : points) {
-            if (sameBlock(point, target))
-                return true;
-        }
-
-        return false;
-    }
-
-    private List<Vector> createCenterPath(List<Vector> controlPoints) {
-        List<Vector> centerPath = new ArrayList<>();
-
-        for (int i = 0; i < controlPoints.size() - 1; i++) {
-            Vector from = controlPoints.get(i);
-            Vector to = controlPoints.get(i + 1);
-
-            appendEightDirectionalLine(centerPath, from, to);
-        }
-
-        return removeOnlyConsecutiveDuplicates(repairGaps(centerPath));
-    }
-
-    private void appendEightDirectionalLine(List<Vector> path, Vector from, Vector to) {
-        int startX = from.getBlockX();
-        int startY = from.getBlockY();
-        int startZ = from.getBlockZ();
-
-        int endX = to.getBlockX();
-        int endY = to.getBlockY();
-        int endZ = to.getBlockZ();
-
-        int dx = endX - startX;
-        int dy = endY - startY;
-        int dz = endZ - startZ;
-
-        int steps = Math.max(Math.abs(dx), Math.abs(dz));
-
-        if (steps == 0) {
-            addPointIfNew(path, new Vector(startX, startY, startZ));
-            return;
-        }
-
-        for (int step = 0; step <= steps; step++) {
-            double t = step / (double) steps;
-
-            int x = (int) Math.round(startX + dx * t);
-            int y = (int) Math.round(startY + dy * t);
-            int z = (int) Math.round(startZ + dz * t);
-
-            addPointIfNew(path, new Vector(x, y, z));
-        }
-    }
-
-    private List<Vector> repairGaps(List<Vector> path) {
-        if (path.size() < 2)
-            return path;
-
-        List<Vector> repaired = new ArrayList<>();
-        repaired.add(path.get(0));
-
-        for (int i = 1; i < path.size(); i++) {
-            Vector previous = repaired.get(repaired.size() - 1);
-            Vector current = path.get(i);
-
-            if (getChebyshevDistance(previous, current) <= 1) {
-                addPointIfNew(repaired, current);
-                continue;
-            }
-
-            appendEightDirectionalLine(repaired, previous, current);
-        }
-
-        return repaired;
-    }
-
-    private int getChebyshevDistance(Vector a, Vector b) {
-        int dx = Math.abs(a.getBlockX() - b.getBlockX());
-        int dz = Math.abs(a.getBlockZ() - b.getBlockZ());
-
-        return Math.max(dx, dz);
-    }
-
-    private void addPointIfNew(List<Vector> path, Vector point) {
-        if (path.isEmpty()) {
-            path.add(point);
-            return;
-        }
-
-        Vector last = path.get(path.size() - 1);
-
-        if (!sameBlock(last, point))
-            path.add(point);
-    }
-
-    private List<Vector> removeOnlyConsecutiveDuplicates(List<Vector> path) {
-        List<Vector> result = new ArrayList<>();
-
-        for (Vector point : path)
-            addPointIfNew(result, point);
-
-        return result;
-    }
-
-    private Vector toBlockVector(Vector vector) {
-        return new Vector(
-                Math.round(vector.getX()),
-                Math.round(vector.getY()),
-                Math.round(vector.getZ())
-        );
-    }
-
-    private boolean sameBlock(Vector a, Vector b) {
-        return a.getBlockX() == b.getBlockX()
-                && a.getBlockY() == b.getBlockY()
-                && a.getBlockZ() == b.getBlockZ();
-    }
-
-    private void placeSampleTrack(List<Vector> centerPath) {
         Map<BlockVector3, BlockData> blockMap = new LinkedHashMap<>();
         LinkedHashSet<BlockVector3> centerPositions = new LinkedHashSet<>();
         Map<BlockVector3, SideBlock> sideBlocks = new LinkedHashMap<>();
